+++ NEWS.md
@@ -2,7 +2,7 @@
 
 # development version
 
-- Add Peng 2025 meropenem ([doi:10.2147/DDDT.S489603](https://doi.org/10.2147/DDDT.S489603)) -- Chinese critically ill adults on continuous renal replacement therapy.
+- Add Chen 2025 ceftazidime and avibactam ([doi:10.2147/IDR.S495279](https://doi.org/10.2147/IDR.S495279)) -- critically and non-critically ill Chinese adults with carbapenem-resistant *Klebsiella pneumoniae* infection.
 
 - Add Sager 2023 sotrovimab population PK and COVID-19 progression exposure-response models ([doi:10.1002/psp4.12958](https://doi.org/10.1002/psp4.12958)) -- non-hospitalized adults and adolescents with mild-to-moderate COVID-19 at high risk of progression, plus healthy volunteers.